Renamed file
Finally renamed nurse to nurse-responsibilities so it matches the other files and isn't the same name as the behavior tree file.
